+++ jgclark.Dashboard/src/react/reducers/actionTypes.js
@@ -6,5 +6,6 @@ export const DASHBOARD_ACTIONS = {
 
 export const PERSPECTIVE_ACTIONS = {
   SET_PERSPECTIVE_SETTINGS: 'SET_PERSPECTIVE_SETTINGS',
+  SET_ACTIVE_PERSPECTIVE: 'SET_ACTIVE_PERSPECTIVE',
   // Add other perspective-related actions here (like update etc because all we have is overwrite)
 }